+++ src/main/java/com/rmm/services/serverapp/model/MonthlyBilling.java
@@ -1,5 +1,6 @@
 package com.rmm.services.serverapp.model;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 /**
@@ -9,34 +10,34 @@ public class MonthlyBilling {
     /**
      * Represents the total bill amount.
      */
-    private Double total;
+    private BigDecimal total;
 
     /**
      * Represents a summary description of the bill amount.
      */
-    private Map<String, Double> summary;
+    private Map<String, BigDecimal> summary;
 
     /**
      * Creates a new instance of {@link MonthlyBilling}.
      */
-    public MonthlyBilling(Double total, Map<String, Double> summary) {
+    public MonthlyBilling(BigDecimal total, Map<String, BigDecimal> summary) {
         this.total = total;
         this.summary = summary;
     }
 
-    public Double getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
 
-    public void setTotal(Double total) {
+    public void setTotal(BigDecimal total) {
         this.total = total;
     }
 
-    public Map<String, Double> getSummary() {
+    public Map<String, BigDecimal> getSummary() {
         return summary;
     }
 
-    public void setSummary(Map<String, Double> summary) {
+    public void setSummary(Map<String, BigDecimal> summary) {
         this.summary = summary;
     }
 }
